+++ packages/magento-newsletter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 9.0.0-canary.56
+
+## 9.0.0-canary.55
+
+## 9.0.0-canary.54
+
+## 8.1.0-canary.53
+
 ## 8.1.0-canary.52
 
 ### Minor Changes